Renamed Capistrano::Systemd::MultiService to Capistrano::Systemd::MultiService::SystemService
